Start to support logins without cookies.
This uses the state parameter to link up an original request with the final one. This means there doesn't need to be a cookie used between the first and last request. However this does expose a potential CSRF attack as the last request doesn't need any state on the client.

There's basic support to mitigate this with an IP restriction that requires that the first and last request are made from the same IP address, however this isn't ideal as it's perfectly possible for requests to come from different IPs (for example carrier grade NAT, mobile networks where access it temporarily lost, other proxy solutions).

At the moment when configuring LTI 1.3 you can choose to use the session based support or the state parameter.

Fixes: #2