Improve the HTML templates.
This extracts the templates out to files loaded at initialisation time. We just read them into memory at startup, this is to avoid binding the library to a particular templating technology.

I18N is missing as well. But it's expected that any popular application would have it's own custom HTML templates for this.