Support the JWT having a kid
When Canvas validates a JWT against a developer key that is using a JWKS URL it requires that the JWT have a `kid` field in it's header that matches one of the keys in the JWKS file (even if there is only one).

When the JWK is embedded in the developer key it doesn't require a `kid`. This change looks up a key ID and if we have one adds it to the JWT header we are building.