AB#64981 Add back support for mobile apps.
We assume that things will work with a cookie based login if the platform storage specification isn't available. This means that mobile users will see a message the first time they login (as we try to use the platform service) but then after that they will get the cookie set and will just get redirected.