AB#64981 Support mobile on 1st login
It was only allowing logins to work on the second login as although we set a cookie on step 3 of the login to say cookie based logins worked we only checked for cookies in the incomming request when looking at how to redirect users.

Now we set a request attribute as well as a cookie to mark that the user has a browser that supports cookies and we then as well as checking for the cookie check for the request attribute.